requirements.txt
langchain
langchain-community
langchain-groq
langchain-openai
langchain-huggingface
sentence-transformers
qdrant-client
fastapi
uvicorn
python-dotenv
pandas
pymupdf
docling
pyjwt
pydantic
pydantic-settings
ragas
pytest
requests
fastembed
nemoguardrails
presidio-analyzer
presidio-anonymizer
spacy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
httpx
prometheus-fastapi-instrumentator
ragas
datasets
pypdf
flashrank
pymupdf
